new update commit

+++ programming_paradigm/robust_division_calculator.py
@@ -1,7 +1,7 @@
 def safe_divide(numerator, denominator):
     try:
-        num = int(numerator)
-        den = int(denominator)
+        num = float(numerator)
+        den = float(denominator)
 
         try:
             result = num / den
